requirements.txt
opencv-python==4.7.0.72
numpy==1.24.0
pytesseract==0.3.10
Pillow==9.5.0
